Big fix

+++ content/graph/GlobalMinCut.h
@@ -11,21 +11,19 @@
 
 pair<int, vi> globalMinCut(vector<vi> mat) {
 	pair<int, vi> best = {INT_MAX, {}};
-	int n = SZ(mat);
-	vector<vi> co(n);
-	FOR(i,0,n) co[i] = {i};
-	FOR(ph,1,n) {
-		vi w = mat[0];
-		size_t s = 0, t = 0;
+	int n = SZ(mat); vector<vi> co(n);
+	FOR(i, 0, n) co[i] = {i};
+	FOR(ph, 1, n) {
+		vi w = mat[0]; int s = 0, t = 0;
 		FOR(it,0,n-ph) { // O(V^2) -> O(E log V) with prio. queue
 			w[t] = INT_MIN;
 			s = t, t = max_element(all(w)) - w.begin();
-			FOR(i,0,n) w[i] += mat[t][i];
+			FOR(i, 0, n) w[i] += mat[t][i];
 		}
 		best = min(best, {w[t] - mat[t][t], co[t]});
 		co[s].insert(co[s].end(), all(co[t]));
-		FOR(i,0,n) mat[s][i] += mat[t][i];
-		FOR(i,0,n) mat[i][s] = mat[s][i];
+		FOR(i, 0, n) mat[s][i] += mat[t][i];
+		FOR(i, 0, n) mat[i][s] = mat[s][i];
 		mat[0][t] = INT_MIN;
 	}
 	return best;
